requirements.txt
django-bootstrap4==0.0.8
django-guardian==2.0.0
django-simple-captcha==0.5.11
djongo==1.2.32
social-auth-app-django==3.1.0
python-gitlab==1.8.0
